Migrate CTS-related code into a separate app
This avoids extra dependencies and the crashing celery task on
systems which do not use CTS.
