+++ conf/celeryconfig.py
@@ -1,9 +1,11 @@
+import datetime
 import logging
 import os
 
 from celery.schedules import crontab
 
 from chronam.settings import *  # NOQA
+from chronam.settings import INSTALLED_APPS
 
 APP_DIR = os.path.dirname(os.path.dirname(os.path.realpath(__file__)))
 
@@ -14,11 +16,6 @@ CELERYD_LOG_LEVEL = logging.INFO
 CELERYD_CONCURRENCY = 2
 
 CELERYBEAT_SCHEDULE = {
-    "poll_cts": {
-        "task": "chronam.core.tasks.poll_cts",
-        "schedule": datetime.timedelta(hours=4),
-        "args": ()
-    },
     "load_essays": {
         "task": "chronam.core.tasks.load_essays",
         "schedule": crontab(hour=0, minute=0),
@@ -29,6 +26,13 @@ CELERYBEAT_SCHEDULE = {
         "schedule": crontab(hour=5, minute=0),
     },
 }
+
+if 'chronam.loc_cts' in INSTALLED_APPS:
+    CELERYBEAT_SCHEDULE["poll_cts"] = {
+        "task": "chronam.core.tasks.poll_cts",
+        "schedule": datetime.timedelta(hours=4),
+        "args": ()
+    }
 
 CELERYBEAT_LOG_FILE = os.path.join("/var/log/celery", "celerybeat.log")
 CELERYBEAT_LOG_LEVEL = logging.INFO
