+++ core/tests/__init__.py
@@ -1,20 +1,18 @@
 import logging
 
-from title_loader_tests import *
-from holding_loader_tests import *
-from title_pull_tests import *
+from api_tests import *
 from batch_loader_tests import *
 from browse_tests import *
-from ocr_extractor_tests import *
 from essay_loader_tests import *
+from holding_loader_tests import *
 from index_tests import *
 from jp2_tests import *
-from api_tests import *
 from json_tests import *
+from ocr_dump_tests import *
+from ocr_extractor_tests import *
 from rdf_tests import *
 from system_tests import *
-from cts_tests import *
-from ocr_dump_tests import *
-
+from title_loader_tests import *
+from title_pull_tests import *
 
 logging.basicConfig(filename="test.log", level=logging.DEBUG)
