+++ loc_cts/tests/__init__.py
@@ -1,0 +1,1 @@
+from cts_tests import *
