+++ settings_loc.py
@@ -18,3 +18,10 @@ DEBUG = False
 IS_PRODUCTION = True
 OMNITURE_SCRIPT = "https://cdn.loc.gov/js/global/metrics/sc/s_code.js"
 SHARETOOL_URL = "https://cdn.loc.gov/sites/chronicling-america.js"
+
+CTS_USERNAME = "username"
+CTS_PASSWORD = "password"
+CTS_PROJECT_ID = "ndnp"
+CTS_QUEUE = "ndnpingestqueue"
+CTS_SERVICE_TYPE = "ingest.NdnpIngest.ingest"
+CTS_URL = "https://cts.loc.gov/transfer/"
